# Optimize `array!` DSL