This optimizes the `array!` DSL to run faster with less memory usage. The PR may look a bit bigger than it actually is, but a summary of the changes:
- Save on a call to `one?`. This shows up as a hotspot for us, and it isn't actually needed. The intent here was to check if a single argument was provided to `array!` to determine whether or not a partial should be rendered. It's actually just faster to check if the first argument is a `Hash` than it is to call out to `one?` first, because the latter is an O(n) operation.
- Save on calls to `::Kernel.block_given?`. Because `Jbuilder` is a `BasicObject`, it does not have direct access to `block_given?`, so it must explicitly call out to the `Kernel` module instead. This is actually slower than simply explicitly checking if the `block` argument exists.
- Saves on memory allocations originally caused by internal calls to `array!`, which splats `*args` into an allocated `Array` each time. The PR introduces `_array` to be used internally, which saves on this allocation. This is similar to what was done in #7 for `extract!`.
- Introduces a frozen `EMPTY_ARRAY` to save on allocating a new empty array (ie. `[]`) each time an empty collections are rendered.

The leanest way to benchmark the `array!` DSL is with

```ruby
json.array!(nil)
```

This gets the main changes running the most frequently under the benchmark without it being diluted by other things. Results look good so far!

```
ruby 3.4.4 (2025-05-14 revision a38531fd3f) +YJIT +PRISM [arm64-darwin24]
Warming up --------------------------------------
              before   491.189k i/100ms
               after   727.930k i/100ms
Calculating -------------------------------------
              before      5.124M (±15.6%) i/s  (195.18 ns/i) -     24.559M in   5.025914s
               after      8.195M (± 7.9%) i/s  (122.03 ns/i) -     40.764M in   5.043596s

Comparison:
               after:  8194919.7 i/s
              before:  5123561.6 i/s - 1.60x  slower
```

```
Calculating -------------------------------------
              before   832.000  memsize (   520.000  retained)
                         8.000  objects (     4.000  retained)
                         0.000  strings (     0.000  retained)
               after   640.000  memsize (   520.000  retained)
                         7.000  objects (     4.000  retained)
                         0.000  strings (     0.000  retained)

Comparison:
               after:        640 allocated
              before:        832 allocated - 1.30x more
```

There are a few other spots that are now also taking advantage of these optimizations. Those have been annotated below.

Please note that all benchmarks are being compared to our fork's `main` branch, so they capture the differences introduced by this PR. If you like, I can also include benchmarks against `upstream_main` to show how _all_ of our optimizations compare to stock `jbuilder` thus far.

(Still poking around the benchmarks. I will include them in the PR soon.)